requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic[email]==2.5.0
python-dotenv==1.0.0
jinja2==3.1.2
slowapi==0.1.9

